web/requirements.txt
django==2.0.4
redis==2.10.6
django-redis==4.9.0
gunicorn==19.8.0
psycopg2==2.7.4
pillow==5.1.0
setuptools
django-imagekit
django-import-export==1.0.1
Unidecode==1.0.22
parse==1.8.4
